struct pocl_argument: use size_t

+++ lib/CL/pocl_cl.h
@@ -327,11 +327,11 @@ extern pocl_obj_id_t last_object_id;
 
 /* Stores kernel argument data defined by clSetKernelArg(). */
 typedef struct pocl_argument {
-  uint64_t size;
+  size_t size;
   /* The "offset" is used to simplify subbuffer handling.
    * At enqueue time, subbuffers are converted to buffers + offset into them.
    */
-  uint64_t offset;
+  size_t offset;
   void *value;
   /* 1 if this argument has been set by clSetKernelArg. */
   char is_set;
